refactor: rename AISDKClient.ts to ai-sdk-client.ts for kebab-case consistency
Rename the file to use kebab-case to match the project's file naming
convention. Update all import statements that reference the old filename.
